# public/openapi.yaml
openapi: 3.0.3
info:
  title: TruthSource API
  description: |
    The TruthSource API enables seamless integration with your B2B e-commerce data accuracy platform.
    Synchronize inventory, pricing, and delivery data between ERPs and e-commerce platforms in real-time.

    ## Authentication
    All API requests require authentication using an API key. Include your API key in the `X-API-Key` header:
    ```
    X-API-Key: your_api_key_here
    ```

    ## Rate Limiting
    - Standard tier: 1,000 requests per hour
    - Professional tier: 10,000 requests per hour
    - Enterprise tier: Unlimited (fair use policy applies)

    ## Base URL
    All API endpoints are relative to: `https://api.truthsource.io/v1`
  version: 1.0.0
  contact:
    name: TruthSource Support
    email: support@truthsource.io
    url: https://truthsource.io/support
  license:
    name: Proprietary
    url: https://truthsource.io/terms
servers:
  - url: https://api.truthsource.io/v1
    description: Production server
  - url: https://sandbox.api.truthsource.io/v1
    description: Sandbox server for testing
...
security:
  - ApiKeyAuth: []
paths:
...
  /auth/api-keys:
    post:
      tags: [Authentication]
      summary: Create new API key
      description: Creates a new API key for accessing the TruthSource API
      operationId: createApiKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, permissions]
              properties:
                name:
                  type: string
                  description: Descriptive name for the API key
                  example: 'Production Integration'
                permissions:
                  type: array
                  items:
                    type: string
                    enum:
                      [
                        read:inventory,
                        write:inventory,
                        read:pricing,
                        write:pricing,
                        read:orders,
                        write:orders,
                      ]
                  example: ['read:inventory', 'write:inventory']
                expiresAt:
                  type: string
                  format: date-time
                  description: Optional expiration date for the API key
      responses:
        '201':
          description: API key created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    get:
      tags: [Authentication]
      summary: List API keys
      description: Retrieves a list of all API keys for the authenticated organization
      operationId: listApiKeys
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKey'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
  /inventory:
    get:
      tags: [Inventory]
      summary: List inventory items
      description: Retrieves a paginated list of inventory items with optional filtering
      operationId: listInventory
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: warehouseId
          in: query
          description: Filter by warehouse ID
          schema:
            type: string
            format: uuid
        - name: productId
          in: query
          description: Filter by product ID
          schema:
            type: string
            format: uuid
        - name: lowStock
          in: query
          description: Only return items below reorder point
          schema:
            type: boolean
      responses:
        '200':
          description: List of inventory items
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/InventoryItem'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags: [Inventory]
      summary: Create inventory item
      description: Creates a new inventory item or updates existing if product/warehouse combination exists
      operationId: createInventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInventoryRequest'
      responses:
        '201':
          description: Inventory item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItem'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
    ApiKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        key:
          type: string
          description: Only returned on creation
        permissions:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
    InventoryItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        warehouseId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 0
        reservedQuantity:
          type: integer
          minimum: 0
        availableQuantity:
          type: integer
          minimum: 0
        reorderPoint:
          type: integer
          nullable: true
        reorderQuantity:
          type: integer
          nullable: true
        lastSync:
          type: string
          format: date-time
        syncStatus:
          type: string
          enum: [synced, pending, error]
        product:
          $ref: '#/components/schemas/Product'
        warehouse:
          $ref: '#/components/schemas/Warehouse'
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        categoryId:
          type: string
          format: uuid
          nullable: true
        unitOfMeasure:
          type: string
        weight:
          type: number
          nullable: true
        dimensions:
          type: object
          nullable: true
          properties:
            length:
              type: number
            width:
              type: number
            height:
              type: number
            unit:
              type: string
              enum: [cm, in]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Warehouse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        address:
          type: object
          properties:
            street:
              type: string
            city:
              type: string
            state:
              type: string
            postalCode:
              type: string
            country:
              type: string
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              nullable: true
    CreateInventoryRequest:
      type: object
      required: [productId, warehouseId, quantity]
      properties:
        productId:
          type: string
          format: uuid
        warehouseId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 0
        reorderPoint:
          type: integer
          minimum: 0
          nullable: true
        reorderQuantity:
          type: integer
          minimum: 1
          nullable: true
...
  responses:
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: BAD_REQUEST
              message: Invalid request parameters
              details:
                field: quantity
                reason: Must be a positive integer
    Unauthorized:
      description: Unauthorized - Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHORIZED
              message: Invalid or missing API key
...